Refactor Push Messaging&Snackbar Service
Foreground+Background notification works and refactored code
