+++ purple-point-front/src/app/services/messaging/messaging.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MessagingService } from './messaging.service';
+
+describe('MessagingService', () => {
+  let service: MessagingService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(MessagingService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
